Cleaning up.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2060 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUv++/CanvasPane.cc
@@ -38,7 +38,7 @@ namespace v
 /************************************************************************
 *  class CanvasPane							*
 ************************************************************************/
-CanvasPane::CanvasPane(Window& parentWin, u_int devWidth, u_int devHeight)
+CanvasPane::CanvasPane(Window& parentWin, size_t devWidth, size_t devHeight)
     :Pane(parentWin),
      _widget(XtVaCreateManagedWidget("TUvCanvasPane",
 				     vViewportWidgetClass,  // widget class
